Final Push for Project B
reorganized Algorithm code into package. Cleaned Main Class for other usage. Added adj() method for nodes in order to find adjacent nodes quicker.
