+++ src/main/java/edu/wpi/teamname/algorithms/edge.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamname;
+package edu.wpi.teamname.algorithms;
 
 public class edge {
   public int edgeID;
