+++ src/main/java/edu/wpi/teamname/algorithms/node.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamname;
+package edu.wpi.teamname.algorithms;
 
 import java.util.LinkedList;
 
@@ -35,6 +35,15 @@ public class node {
     this.edges = new LinkedList<edge>();
   }
 
+  // simple constructor
+  public node(int _nodeID, double _xcoord, double _ycoord, Graph graph) {
+    this.nodeID = _nodeID;
+    this.xcoord = _xcoord;
+    this.ycoord = _ycoord;
+    this.edges = new LinkedList<edge>();
+    graph.addNode(this);
+  }
+
   // For AStar
   public void setDistToGoal(node destination) {
     double dx = Math.abs(this.xcoord - destination.xcoord);
@@ -45,12 +54,11 @@ public class node {
     this.distanceToGoal = distance;
   }
 
-  // simple constructor
-  public node(int _nodeID, double _xcoord, double _ycoord, Graph graph) {
-    this.nodeID = _nodeID;
-    this.xcoord = _xcoord;
-    this.ycoord = _ycoord;
-    this.edges = new LinkedList<edge>();
-    graph.addNode(this);
+  public LinkedList<node> adjNodes() {
+    LinkedList<node> result = new LinkedList<>();
+    for (edge e : this.edges) {
+      result.add(e.endNode);
+    }
+    return result;
   }
 }
